+++ meongcare/src/main/java/com/meongcare/domain/member/domain/entity/Member.java
@@ -1,4 +1,4 @@
-package com.meongcare.domain.auth.domain.entity;
+package com.meongcare.domain.member.domain.entity;
 
 import lombok.AccessLevel;
 import lombok.Builder;
