+++ meongcare/src/main/java/com/meongcare/domain/member/domain/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.meongcare.domain.auth.domain.repository;
+package com.meongcare.domain.member.domain.repository;
 
-import com.meongcare.domain.auth.domain.entity.Member;
+import com.meongcare.domain.member.domain.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
